colisión del rayo con los jefes arreglado

+++ Assets/Scripts/APIScripts/ScoreResponse.cs
@@ -3,5 +3,5 @@
 public class ScoreResponse
 {
     public string name;
-    public float puntuacion;
+    public int puntuacion;
 }
